backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
bcrypt==4.2.1
python-slugify==8.0.1
google-genai==1.58.0
